move utils to server
