+++ lxl-web/src/hooks.server.ts
@@ -7,10 +7,10 @@ import type { QualifierSuggestion2 } from '$lib/types/search';
 import type { Site } from '$lib/types/site';
 import { DebugFlags, type MyLibrariesType, type UserSettings } from '$lib/types/userSettings';
 import displayWeb from '$lib/assets/json/display-web.json';
-import { DisplayUtil, VocabUtil } from '$lib/utils/xl';
+import { DisplayUtil, VocabUtil } from '$lib/utils/xl.server';
 import { startRefreshLibraries } from '$lib/utils/getLibraries.server';
 import { getSubsetMapping } from '$lib/utils/subsetCache.server';
-import { getQualifierSuggestions } from '$lib/utils/getQualifierSuggestions';
+import { getQualifierSuggestions } from '$lib/utils/getQualifierSuggestions.server';
 
 type QualifierSuggestionsByLocale = Record<keyof typeof Locales, QualifierSuggestion2[]>;
 type Util = [VocabUtil, DisplayUtil, QualifierSuggestionsByLocale];
