+++ lxl-web/src/lib/remotes/searchResult.remote.ts
@@ -4,7 +4,7 @@ import { query, getRequestEvent } from '$app/server';
 import * as v from 'valibot';
 
 import { type ApiError } from '$lib/types/api';
-import { asResult } from '$lib/utils/search';
+import { asResult } from '$lib/utils/search.server';
 import type { PartialCollectionView } from '$lib/types/search';
 import { DebugFlags } from '$lib/types/userSettings';
 import { getSupportedLocale } from '$lib/i18n/locales';
