+++ lxl-web/src/lib/types/relations.ts
@@ -1,0 +1,8 @@
+export type Relation = {
+	qualifierKey: string;
+	label: string;
+	totalItems: number;
+	findUrl: string;
+	previewUrl: string;
+	isLike: boolean;
+};
