+++ lxl-web/src/lib/utils/cslFromMainEntity.test.ts
@@ -1,7 +1,7 @@
 import mainEntity from '$lib/assets/json/test-data/main-entity.json';
 import { describe, it, expect } from 'vitest';
 import { cslFromMainEntity } from './cslFromMainEntity';
-import type { VocabUtil } from './xl';
+import type { VocabUtil } from './xl.server';
 import type { FramedData } from '$lib/types/xl';
 
 const vocabUtil = { getDefinition: (a: string) => a as unknown as FramedData } as VocabUtil;
