+++ lxl-web/src/lib/utils/getQualifierSuggestions.server.ts
@@ -1,5 +1,6 @@
 import { type LocaleCode, otherLocales } from '$lib/i18n/locales';
-import { asArray, type DisplayUtil, toString, VocabUtil } from '$lib/utils/xl';
+import { type DisplayUtil, VocabUtil } from '$lib/utils/xl.server';
+import { asArray, toString } from '$lib/utils/xl';
 import { type FramedData, JsonLd, LensType, Platform } from '$lib/types/xl';
 import { type QualifierSuggestion2, QualifierSuggestionShowIn } from '$lib/types/search';
 import { getUriSlug } from '$lib/utils/http';
