+++ lxl-web/src/lib/utils/relations.server.ts
@@ -1,6 +1,6 @@
 import { env } from '$env/dynamic/private';
 import { error } from '@sveltejs/kit';
-import type { VocabUtil } from '$lib/utils/xl';
+import type { VocabUtil } from '$lib/utils/xl.server';
 import type { PartialCollectionView } from '$lib/types/search';
 import { Base, JsonLd, Platform } from '$lib/types/xl';
 import addDefaultSearchParams from '$lib/utils/addDefaultSearchParams';
@@ -9,15 +9,6 @@ import capitalize from '$lib/utils/capitalize';
 import { type ApiError } from '$lib/types/api.js';
 import getAtPath from '$lib/utils/getAtPath';
 import { getUriSlug } from '$lib/utils/http';
-
-export type Relation = {
-	qualifierKey: string;
-	label: string;
-	totalItems: number;
-	findUrl: string;
-	previewUrl: string;
-	isLike: boolean;
-};
 
 export async function getRelations(
 	resourceId: string | null,
